requirements.txt
crewai==0.111.1
langchain
langchain-openai
openai
supabase
python-dotenv
streamlit